// pharmacy-ui/src/app/services/models/product-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProductRequest {
  description?: string;
  expiryDate?: string;
  image?: string;
  manufacturer?: string;
  medicineId?: number;
  name?: string;
  price?: number;
  stockQuantity?: number;
  type?: 'SYRUP' | 'TABLETS' | 'OINTMENTS';
}